% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raking_formula.R
\name{collect_interaction_vars}
\alias{collect_interaction_vars}
\title{Recursively collect variables from an interaction expression}
\usage{
collect_interaction_vars(expr)
}
\description{
Recursively collect variables from an interaction expression
}
\keyword{internal}
