% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raking_formula.R
\name{create_constraint_term}
\alias{create_constraint_term}
\title{Create a term specification for l2/kl constraints}
\usage{
create_constraint_term(type, expr)
}
\description{
Create a term specification for l2/kl constraints
}
\keyword{internal}
