% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raking_formula.R
\name{create_exact_term}
\alias{create_exact_term}
\title{Create a term specification for exact matching}
\usage{
create_exact_term(expr)
}
\description{
Create a term specification for exact matching
}
\keyword{internal}
