% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raking_formula.R
\name{create_interaction_term}
\alias{create_interaction_term}
\title{Create a term specification for interactions}
\usage{
create_interaction_term(vars)
}
\description{
Create a term specification for interactions
}
\keyword{internal}
