% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raking_formula.R
\name{create_quantile_term}
\alias{create_quantile_term}
\title{Create a term specification for quantile constraints}
\usage{
create_quantile_term(expr, p)
}
\description{
Create a term specification for quantile constraints
}
\keyword{internal}
