% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raking_formula.R
\name{create_term_id}
\alias{create_term_id}
\title{Create a unique identifier for a term}
\usage{
create_term_id(type, expr)
}
\description{
Create a unique identifier for a term
}
\keyword{internal}
