% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raking_formula.R
\name{formula2str}
\alias{formula2str}
\title{Convert a formula to a string representation}
\usage{
formula2str(x)
}
\arguments{
\item{x}{A formula object}
}
\value{
A string representation of the formula
}
\description{
Converts a formula object to a readable string format.
This is an internal helper function used by the print methods.
}
\keyword{internal}
