% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raking_formula.R
\name{get_constraint_handler}
\alias{get_constraint_handler}
\title{Get a constraint handler}
\usage{
get_constraint_handler(type)
}
\arguments{
\item{type}{Type of constraint}
}
\value{
A handler function for the specified constraint type
}
\description{
Retrieves the handler function for a specific constraint type.
}
\keyword{internal}
