% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrake.R
\name{print.regrake}
\alias{print.regrake}
\title{Print method for regrake objects}
\usage{
\method{print}{regrake}(x, ...)
}
\arguments{
\item{x}{A regrake object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the object
}
\description{
Print method for regrake objects
}
