% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raking_formula.R
\name{process_term}
\alias{process_term}
\title{Process a raking term}
\usage{
process_term(term, data)
}
\arguments{
\item{term}{A raking term object}

\item{data}{A data frame}
}
\value{
A processed term specification
}
\description{
Process a raking term
}
\keyword{internal}
