% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizers.R
\name{prox_equality_reg}
\alias{prox_equality_reg}
\title{Proximal operator for equality regularizer}
\usage{
prox_equality_reg(w, lam)
}
\arguments{
\item{w}{Input vector}

\item{lam}{Regularization parameter}
}
\value{
Original vector (identity operation)
}
\description{
Proximal operator for equality regularizer
}
\examples{
prox_equality_reg(c(0.1, 0.2), lam = 0.5)
}
