% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizers.R
\name{prox_sum_squares_reg}
\alias{prox_sum_squares_reg}
\title{Proximal operator for sum squares regularizer}
\usage{
prox_sum_squares_reg(w, lam)
}
\arguments{
\item{w}{Input vector}

\item{lam}{Regularization parameter}
}
\value{
Updated vector minimizing sum squares plus proximal term
}
\description{
Proximal operator for sum squares regularizer
}
