% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrake-package.R
\docType{package}
\name{regrake-package}
\alias{regrake-package}
\title{regrake: Regularized Raking for Survey Weighting}
\description{
A modern implementation of regularized survey raking based on the
mathematical framework from Barratt et al. (2021).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andytimm/regrake}
  \item Report bugs at \url{https://github.com/andytimm/regrake/issues}
}

}
\author{
\strong{Maintainer}: Andy Timm \email{timmandrew1@gmail.com} [copyright holder]

}
\keyword{internal}
