% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrake.R
\name{summary.regrake}
\alias{summary.regrake}
\title{Summary method for regrake objects}
\usage{
\method{summary}{regrake}(object, ...)
}
\arguments{
\item{object}{A regrake object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the object
}
\description{
Summary method for regrake objects
}
