% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raking_formula.R
\name{validate_overlapping_constraints}
\alias{validate_overlapping_constraints}
\title{Validate that there are no invalid overlapping constraints}
\usage{
validate_overlapping_constraints(terms)
}
\description{
Validate that there are no invalid overlapping constraints
}
\keyword{internal}
