% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raking_formula.R
\name{create_constraint_term}
\alias{create_constraint_term}
\title{Create a term specification for l2/kl constraints}
\usage{
create_constraint_term(type, expr, wrapper_name = paste0("rr_", type))
}
\arguments{
\item{type}{Constraint type.}

\item{expr}{Expression to parse.}

\item{wrapper_name}{Optional wrapper function name used for clearer error messages.}
}
\description{
Create a term specification for l2/kl constraints
}
\keyword{internal}
